add remove mechanic

+++ src/towerDefense/entity/bullets/BulletClass.java
@@ -5,13 +5,13 @@ import javafx.scene.layout.Pane;
 import towerDefense.EntityClass;
 
 public abstract class BulletClass extends EntityClass {
-    private double damage;
+    private int damage;
 
-    protected BulletClass(Pane layer, Image image, long tick, double posX, double posY, double rotation, double damage)
+    protected BulletClass(Pane layer, Image image, long tick, double posX, double posY, double rotation, int damage)
     {
         super(layer, image, tick, posX, posY, rotation);
         this.damage = damage;
     }
 
-    public double getDamage() { return this.damage; }
+    public int getDamage() { return this.damage; }
 }
